# Speed up DOM rendering