--- a/client/views/console-cell-view.js
+++ b/client/views/console-cell-view.js
@@ -41,7 +41,7 @@ define([
                     this.model.get("backgroundGreen") + "," +
                     this.model.get("backgroundBlue") + ")";
 
-            this.el.innerHTML = cellCharacter;
+            this.el.textContent = cellCharacter;
             this.el.style.color = rgbForegroundString;
             this.el.style.backgroundColor = rgbBackgroundString;
             return this;