Two small tweaks to make DOM layout computation about 15 times faster on Chrome. (This is in a VM with no GPU acceleration.)

Before patch: 2 typical frames are rendered in 611 ms. DOM layout wastes ~200 ms per frame.

![before](https://user-images.githubusercontent.com/69065091/107865712-bcc87b00-6e26-11eb-8b81-f87dbd727909.png)

After patch: 2 typical frames are rendered in 46 ms. DOM layout takes just 3 ms per frame.

![after](https://user-images.githubusercontent.com/69065091/107865716-c2be5c00-6e26-11eb-9f8b-27ab2b24992e.png)

There is also a gain on Firefox, but not as dramatic: DOM drops from 42% of CPU time to ~35% of CPU time. Firefox feels smoother than Chrome both before and after.