@flend 2021-02-15
Amazing speedup on this unloved part of the codebase, thanks!